fix: make data more dynamic
