+++ src/components/resume/travel-map/cruises/CruiseCharts.tsx
@@ -5,7 +5,7 @@ import sankey from "highcharts/modules/sankey";
 import highchartsAccessibility from "highcharts/modules/accessibility";
 import { useRecoilState } from "recoil";
 import themeAtom from "../../../../recoil/theme-atom";
-import { cruiseData } from "../../../../constants/travel";
+import { cruiseData } from "../../../../constants/cruises";
 
 sankey(Highcharts); // initiate sankey module
 highchartsMore(Highcharts); // if your module is not in node_modules folder
