+++ src/components/resume/travel-map/cruises/LoyaltyCharts.tsx
@@ -2,9 +2,8 @@ import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
 import highchartsAccessibility from "highcharts/modules/accessibility";
 import { useRecoilState } from "recoil";
-import { amber, red, grey, blueGrey } from "@mui/material/colors";
 import themeAtom from "../../../../recoil/theme-atom";
-// import { cruiseData } from "../../../../constants/travel";
+import { loyaltySeries, loyaltyColors } from "../../../../constants/cruises";
 
 highchartsAccessibility(Highcharts); // initiate accessibility module
 
@@ -13,44 +12,26 @@ const LoyaltyCharts = () => {
   const color = theme.mode === "light" ? "black" : "white";
 
   const options: Highcharts.Options = {
-    colors: [
-      red[100],
-      grey[400],
-      amber[500],
-      blueGrey[900],
-      grey[50],
-      "transparent",
-    ],
+    colors: loyaltyColors,
     chart: {
       type: "column",
       inverted: true,
       polar: true,
       backgroundColor: "transparent",
     },
-    title: {
-      text: "Cruise Loyalty",
-      style: { color },
-    },
+    title: { text: "Cruise Loyalty", style: { color } },
     tooltip: {
-      outside: true,
       valueSuffix: "%",
     },
-    pane: {
-      size: "85%",
-      innerSize: "20%",
-      endAngle: 335,
-    },
+    pane: { size: "85%", innerSize: "20%", endAngle: 335 },
+    legend: { itemStyle: { color } },
     xAxis: {
-      tickInterval: 1,
       labels: {
         align: "right",
-        useHTML: true,
-        allowOverlap: true,
         step: 1,
         y: 3,
         style: { color },
       },
-      lineWidth: 0,
       gridLineWidth: 0,
       categories: [
         "Disney 🛳️",
@@ -67,7 +48,7 @@ const LoyaltyCharts = () => {
       endOnTick: true,
       showLastLabel: true,
       gridLineWidth: 0,
-      labels: { style: { color } },
+      labels: { format: "{text}%", style: { color } },
     },
     plotOptions: {
       column: {
@@ -78,45 +59,7 @@ const LoyaltyCharts = () => {
         borderRadius: "50%",
       },
     },
-    series: [
-      {
-        type: "column",
-        name: "None",
-        data: [
-          Math.floor((1 / 30) * 100),
-          Math.floor((3 / 30) * 100),
-          Math.floor((1 / 3) * 100),
-          Math.floor((1 / 30) * 100),
-          Math.floor((1 / 2) * 100),
-        ],
-      },
-      {
-        type: "column",
-        name: "Silver",
-        data: [Math.floor((4 / 30) * 100), 0, 0, 0, 0],
-      },
-      {
-        type: "column",
-        name: "Gold",
-        data: [
-          Math.floor((5 / 30) * 100),
-          Math.floor((1 / 30) * 100),
-          0,
-          Math.floor((3 / 30) * 100),
-          0,
-        ],
-      },
-      {
-        type: "column",
-        name: "Platinum",
-        data: [Math.floor((15 / 30) * 100), 0, 0, 0, 0],
-      },
-      {
-        type: "column",
-        name: "Pearl",
-        data: [Math.floor((5 / 30) * 100), 0, 0, 0, 0],
-      },
-    ],
+    series: loyaltySeries,
   };
 
   return (
