+++ src/components/resume/travel-map/map/CountryTable.tsx
@@ -70,7 +70,7 @@ for (let i = 0; i < len; i += 1) {
 }
 
 // --------------------     Travel Map     -------------------- //
-const TravelMap: React.FC = React.memo(() => (
+const CountryTable = React.memo(() => (
   <Table aria-label="Countries I have visited">
     <TableHead>
       <TableRow>
@@ -90,4 +90,4 @@ const TravelMap: React.FC = React.memo(() => (
   </Table>
 ));
 
-export default TravelMap;
+export default CountryTable;
